# Speed up doctest of `calibrate_denoiser` by ~30 s